refactor: move Login component to Auth directory and update styles
